+++ apps/cowswap-frontend/src/modules/permit/hooks/useOrdersPermitStatus.ts
@@ -1,7 +1,0 @@
-import { useAtomValue } from 'jotai'
-
-import { ordersPermitStatusAtom } from '../state/ordersPermitStatusAtom'
-
-export function useGetOrdersPermitStatus() {
-  return useAtomValue(ordersPermitStatusAtom)
-}