+++ apps/cowswap-frontend/src/modules/volumeFee/state/taxFreeAssetsAtom.ts
@@ -1,0 +1,11 @@
+import { atomWithStorage } from 'jotai/utils'
+
+import { mapSupportedNetworks } from '@cowprotocol/cow-sdk'
+import { PersistentStateByChain } from '@cowprotocol/types'
+
+type TokenId = string
+
+export const taxFreeAssetsAtom = atomWithStorage<PersistentStateByChain<TokenId[][]>>(
+  'taxFreeAssetsAtom:v1',
+  mapSupportedNetworks<TokenId[][]>([]),
+)