+++ apps/explorer/src/components/orders/OrdersUserDetailsTable/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react'
 
+import { Command } from '@cowprotocol/types'
 import { TruncatedText } from '@cowprotocol/ui/pure/TruncatedText'
 
 import { faExchangeAlt } from '@fortawesome/free-solid-svg-icons'
@@ -17,13 +18,27 @@ import { useNetworkId } from 'state/network'
 import styled from 'styled-components/macro'
 import { FormatAmountPrecision, formattedAmount } from 'utils'
 
-import { Order } from 'api/operator'
+import { Order, OrderStatus } from 'api/operator'
 import { getLimitPrice } from 'utils/getLimitPrice'
 
 import { OrderSurplusDisplayStyledByRow } from './OrderSurplusTooltipStyledByRow'
+import { ToggleFilter } from './ToggleFilter'
 
+import { TableState } from '../../../explorer/components/TokensTableWidget/useTable'
 import { SimpleTable, SimpleTableProps } from '../../common/SimpleTable'
 import { StatusLabel } from '../StatusLabel'
+import { UnsignedOrderWarning } from '../UnsignedOrderWarning'
+
+const EXPIRED_CANCELED_STATES: OrderStatus[] = ['cancelled', 'cancelling', 'expired']
+
+function isExpiredOrCanceled(order: Order): boolean {
+  const { executedSellAmount, executedBuyAmount, status } = order
+  // We don't consider an order expired or canceled if it was partially or fully filled
+  if (!executedSellAmount.isZero() || !executedBuyAmount.isZero()) return false
+
+  // Otherwise, return if the order is expired or canceled
+  return EXPIRED_CANCELED_STATES.includes(status)
+}
 
 const tooltip = {
   orderID: 'A unique identifier ID for this order.',
@@ -40,15 +55,68 @@ const Wrapper = styled.div`
 
 export type Props = SimpleTableProps & {
   orders: Order[] | undefined
+  tableState: TableState
+  handleNextPage: Command
   messageWhenEmpty?: string | React.ReactNode
 }
 
 interface RowProps {
   order: Order
   isPriceInverted: boolean
+
+  // TODO: Filter by state using the API. Not available for now, so filtering in the client
+  showCanceledAndExpired: boolean
+  showPreSigning: boolean
 }
 
-const RowOrder: React.FC<RowProps> = ({ order, isPriceInverted }) => {
+const FilterRow = styled.tr`
+  background-color: ${({ theme }) => theme.background};
+
+  @media (max-width: 1155px) {
+    div:first-child {
+      max-width: 90vw;
+    }
+  }
+
+  td {
+    padding: 2rem;
+    text-align: right;
+    padding-right: 10px;
+    max-width: 100%;
+    & > * {
+      margin-left: 10px;
+    }
+  }
+
+  p {
+    word-wrap: break-word;
+    white-space: normal;
+  }
+`
+
+const Filters = styled.div`
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  flex-direction: row;
+  gap: 1rem;
+`
+
+const HiddenOrdersLegend = styled.div`
+  p {
+    text-align: center;
+  }
+
+  a {
+    text-decoration: underline;
+  }
+
+  a:hover {
+    color: ${({ theme }) => theme.textSecondary2};
+  }
+`
+
+const RowOrder: React.FC<RowProps> = ({ order, isPriceInverted, showCanceledAndExpired, showPreSigning }) => {
   const { creationDate, buyToken, buyAmount, sellToken, sellAmount, kind, partiallyFilled, uid, filledPercentage } =
     order
   const [_isPriceInverted, setIsPriceInverted] = useState(isPriceInverted)
@@ -66,6 +134,10 @@ const RowOrder: React.FC<RowProps> = ({ order, isPriceInverted }) => {
   const renderSpinnerWhenNoValue = (textValue: string): React.ReactNode | void => {
     if (textValue === '-') return <Spinner spin size="1x" />
   }
+
+  // Hide the row if the order is canceled, expired or pre-signing
+  if (!showCanceledAndExpired && isExpiredOrCanceled(order)) return null
+  if (!showPreSigning && order.status === 'signing') return null
 
   return (
     <tr key={uid}>
@@ -116,8 +188,19 @@ const RowOrder: React.FC<RowProps> = ({ order, isPriceInverted }) => {
 }
 
 const OrdersUserDetailsTable: React.FC<Props> = (props) => {
-  const { orders, messageWhenEmpty } = props
+  const { orders, messageWhenEmpty, tableState, handleNextPage } = props
   const [isPriceInverted, setIsPriceInverted] = useState(false)
+  const [showCanceledAndExpired, setShowCanceledAndExpired] = useState(false)
+  const [showPreSigning, setShowPreSigning] = useState(false)
+
+  const canceledAndExpiredCount = orders?.filter(isExpiredOrCanceled).length || 0
+  const preSigningCount = orders?.filter((order) => order.status === 'signing').length || 0
+  const showFilter = canceledAndExpiredCount > 0 || preSigningCount > 0
+
+  const hiddenOrdersCount =
+    (showPreSigning ? 0 : preSigningCount) + (showCanceledAndExpired ? 0 : canceledAndExpiredCount)
+
+  const areOrdersAllHidden = orders?.length === hiddenOrdersCount
 
   const invertLimitPrice = (): void => {
     setIsPriceInverted((previousValue) => !previousValue)
@@ -130,30 +213,101 @@ const OrdersUserDetailsTable: React.FC<Props> = (props) => {
   return (
     <SimpleTable
       header={
-        <tr>
-          <th>
-            <span>
-              Order ID <HelpTooltip tooltip={tooltip.orderID} />
-            </span>
-          </th>
-          <th>Type</th>
-          <th>Sell amount</th>
-          <th>Buy amount</th>
-          <th>
-            <span>
-              Limit price <Icon icon={faExchangeAlt} onClick={invertLimitPrice} />
-            </span>
-          </th>
-          <th>Surplus</th>
-          <th>Created</th>
-          <th>Status</th>
-        </tr>
+        <>
+          {!areOrdersAllHidden && (
+            <tr>
+              <th>
+                <span>
+                  Order ID <HelpTooltip tooltip={tooltip.orderID} />
+                </span>
+              </th>
+              <th>Type</th>
+              <th>Sell amount</th>
+              <th>Buy amount</th>
+              <th>
+                <span>
+                  Limit price <Icon icon={faExchangeAlt} onClick={invertLimitPrice} />
+                </span>
+              </th>
+              <th>Surplus</th>
+              <th>Created</th>
+              <th>Status</th>
+            </tr>
+          )}
+          {showPreSigning && (
+            <FilterRow>
+              <td colSpan={8}>
+                <div>
+                  <UnsignedOrderWarning />
+                </div>
+              </td>
+            </FilterRow>
+          )}
+        </>
       }
       body={
         <>
-          {orders.map((item) => (
-            <RowOrder key={item.uid} order={item} isPriceInverted={isPriceInverted} />
-          ))}
+          {!areOrdersAllHidden &&
+            orders.map((item) => (
+              <RowOrder
+                key={item.uid}
+                order={item}
+                isPriceInverted={isPriceInverted}
+                showCanceledAndExpired={showCanceledAndExpired}
+                showPreSigning={showPreSigning}
+              />
+            ))}
+
+          {showFilter && (
+            <FilterRow>
+              <td colSpan={8}>
+                <div>
+                  <HiddenOrdersLegend>
+                    {hiddenOrdersCount > 0 ? (
+                      <>
+                        <p>
+                          Showing {orders.length - hiddenOrdersCount} out of {orders.length} orders for the current
+                          page.
+                        </p>
+                        <p>
+                          {hiddenOrdersCount} orders are hidden, you can make them visible using the filters below
+                          {tableState.hasNextPage ? (
+                            <span>
+                              , or go to&nbsp;<a onClick={handleNextPage}>next page</a>&nbsp;for more orders.
+                            </span>
+                          ) : (
+                            '.'
+                          )}
+                        </p>
+                      </>
+                    ) : (
+                      <p>Showing all {orders.length} orders for the current page.</p>
+                    )}
+                  </HiddenOrdersLegend>
+                  <Filters>
+                    {canceledAndExpiredCount > 0 && (
+                      <ToggleFilter
+                        checked={showCanceledAndExpired}
+                        onChange={() => setShowCanceledAndExpired((previousValue) => !previousValue)}
+                        label={(showCanceledAndExpired ? 'Hide' : 'Show') + ' canceled/expired'}
+                        count={canceledAndExpiredCount}
+                      />
+                    )}
+                    {preSigningCount > 0 && (
+                      <>
+                        <ToggleFilter
+                          checked={showPreSigning}
+                          onChange={() => setShowPreSigning((previousValue) => !previousValue)}
+                          label={(showPreSigning ? 'Hide' : 'Show') + ' unsigned'}
+                          count={preSigningCount}
+                        />
+                      </>
+                    )}
+                  </Filters>
+                </div>
+              </td>
+            </FilterRow>
+          )}
         </>
       }
     />