+++ libs/common-hooks/src/useAvailableChains.ts
@@ -3,8 +3,6 @@ import { useMemo } from 'react'
 import { getAvailableChains } from '@cowprotocol/common-utils'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 
-import { useFeatureFlags } from './useFeatureFlags'
-
 /**
  * Hook to get a list of SupportedChainId currently available/enabled
  *
@@ -14,12 +12,12 @@ import { useFeatureFlags } from './useFeatureFlags'
  */
 export function useAvailableChains(): SupportedChainId[] {
   // 1. Load feature flag for chain being enabled
-  const { isBaseEnabled } = useFeatureFlags()
+  // const { isBaseEnabled } = useFeatureFlags()
 
   return useMemo(
     // 2. Conditionally build a list of chain ids to exclude
-    // () => getAvailableChains(isArbitrumOneEnabled ? undefined : [SupportedChainId.ARBITRUM_ONE]),
-    () => getAvailableChains(isBaseEnabled ? undefined : [SupportedChainId.BASE]),
-    [isBaseEnabled],
+    // () => getAvailableChains(isBaseEnabled ? undefined : [SupportedChainId.BASE]),  <-- example usage, kept for reference
+    () => getAvailableChains(),
+    [],
   )
 }