+++ libs/core/src/cms/utils/querySerializer.ts
@@ -1,0 +1,5 @@
+import qs from 'qs'
+
+export const querySerializer = (params: any) => {
+  return qs.stringify(params, { encodeValuesOnly: true, arrayFormat: 'brackets' })
+}